+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/nms/NmsAccessor.kt
@@ -1,4 +1,4 @@
-package io.github.pylonmc.pylon.core.i18n.packet
+package io.github.pylonmc.pylon.core.nms
 
 import io.github.pylonmc.pylon.core.i18n.PlayerTranslationHandler
 import org.bukkit.entity.Player
@@ -6,7 +6,7 @@ import org.jetbrains.annotations.ApiStatus
 
 @ApiStatus.Internal
 @ApiStatus.NonExtendable
-interface PacketManager {
+interface NmsAccessor {
 
     fun register(player: Player, handler: PlayerTranslationHandler)
 
@@ -15,8 +15,8 @@ interface PacketManager {
     fun resendInventory(player: Player)
 
     companion object {
-        val instance = Class.forName("io.github.pylonmc.pylon.core.i18n.packet.PacketManagerImpl")
+        val instance = Class.forName("io.github.pylonmc.pylon.core.nms.NmsAccessorImpl")
             .getDeclaredField("INSTANCE")
-            .get(null) as PacketManager
+            .get(null) as NmsAccessor
     }
 }
